requirements.txt
requests~=2.28.1
web3~=5.31.3
pyevmasm~=0.2.3
difflib~=3.2
pycryptodome~=3.15.0
